requirements.txt
discord.py
python-dotenv
google-generativeai
huggingface_hub
Flask
requests